ATEAM-7039: RFE- USBank - HashiCorp Vault plugin for CAGW must be compatible with Kubernetes Cert-Manager
The CA configuration concept has been altered to a configuration role concept. Configuration roles are now associated with a CA and optionally a profile.  If the CA name isn't specified when configuring the role, the role name is used as the CA name. If a profile id is not specified the role is considered to be for all profiles of that CA and a profile id must be specified in the issue or sign request body.
Profiles can now be configured at config/{roleName}/profile if the specified role configuration contains a profile.
Pre-existing CA configurations should continue to work. Their names will be used as the CA name and they will be considered for all profiles.
Fixed error message when invalid profile ID is specified and the config cannot be found.
Fixed some more detailed logging.
No longer using a hardcoded password for P12; now a random one is generated.